Spun live collaboration code off into a separate server
I order for the live update code to work with multiple instances of
tornado, there can only be one server that handles the collaboration.
The code is currently written this way. So I had two options: run only
one instance of tornado for the entire site, or spin off the updates
into a separate server. I obviously chose the latter. However, this
relies on nginx to properly forward /updates to live.py running on port
9300, so I'm committing this to the production branch, and not the main
one. This now allows us to run any amount of main tornado instances we
wish, and then one instance of live.py. Communication between the two is
done via AyncHTTPClient().